+++ map_test.go
@@ -1,0 +1,29 @@
+package tilepix_test
+
+import (
+	"testing"
+
+	"github.com/bcvery1/tilepix"
+)
+
+func TestGetLayerByName(t *testing.T) {
+	m, err := tilepix.ReadFile("testdata/poly.tmx")
+	if err != nil {
+		t.Fatal(err)
+	}
+	layer := m.GetLayerByName("Tile Layer 1")
+	if layer.Name != "Tile Layer 1" {
+		t.Error("error get layer")
+	}
+}
+
+func TestGetObjectLayerByName(t *testing.T) {
+	m, err := tilepix.ReadFile("testdata/poly.tmx")
+	if err != nil {
+		t.Fatal(err)
+	}
+	layer := m.GetObjectLayerByName("Object Layer 1")
+	if layer.Name != "Object Layer 1" {
+		t.Error("error get object layer")
+	}
+}
